Entidad Mayorista - Atributo Cuit Int por Long

+++ Back-Ecommerce/Ecommerce-Perfumes/Application/Models/Response/MayoristaResponse.cs
@@ -11,7 +11,7 @@ namespace Application.Models.Response
         public int Id { get; set; }
         public string FirstName { get; set; } = string.Empty;
         public string LastName { get; set; } = string.Empty;
-        public int CUIT { get; set; }
+        public long CUIT { get; set; }
         public string Categoria { get; set; } = string.Empty;
         public string NameAccount { get; set; } = string.Empty;
         public string Password { get; set; } = string.Empty;
