+++ Back-Ecommerce/Ecommerce-Perfumes/Domain/Entities/Mayorista.cs
@@ -3,7 +3,7 @@
     public class Mayorista : User
     {
         public string Categoria { get; set; } = string.Empty;
-        public int CUIT { get; set; }
+        public long CUIT { get; set; }
         public Mayorista()
         {
             TypeUser = "Mayorista";
